changed

+++ tasks/894_B.cpp
@@ -1,15 +1,15 @@
 #include <iostream>
 #include <vector>
 
-std::vector<int> input(int size);
-void original_cons(int n);
+std::vector<int> input(long unsigned int size);
+void original_cons(long unsigned int n);
 
 int main() {
 
-  int t, n;
+  long unsigned int t, n;
 
   std::cin >> t;
-  for (int i = 0; i < t; ++i) {
+  for (long unsigned int i = 0; i < t; ++i) {
     std::cin >> n;
     original_cons(n);
     if (i != t - 1)
@@ -18,23 +18,23 @@ int main() {
 
   return 0;
 }
-std::vector<int> input(int size) {
+std::vector<int> input(long unsigned int size) {
   std::vector<int> res(size);
-  for (int i = 0; i < size; ++i) {
+  for (long unsigned int i = 0; i < size; ++i) {
     std::cin >> res[i];
   }
 
   return res;
 }
 
-void original_cons(int n) {
+void original_cons(long unsigned int n) {
 
   std::vector<int> b = input(n);
   std::vector<int> res(2 * n);
 
   res[0] = b[0];
-  int k = 0;
-  for (int i = 1; i < n; i++) {
+  long unsigned int k = 0;
+  for (long unsigned int i = 1; i < n; i++) {
     if (b[i] >= b[i - 1])
       res[i + k] = b[i];
     else {
@@ -45,7 +45,7 @@ void original_cons(int n) {
   }
   std::cout << n + k << "\n";
 
-  for (int i = 0; i < n + k; i++) {
+  for (long unsigned int i = 0; i < n + k; i++) {
     std::cout << res[i] << " ";
   }
 }
