+++ tasks/894_C.cpp
@@ -1,15 +1,15 @@
 #include <iostream>
 #include <vector>
 
-std::vector<int> input(int size);
-void simmetry(int n);
+std::vector<long unsigned int> input(long unsigned int size);
+void simmetry(long unsigned int n);
 
 int main() {
 
-  int t, n;
+  long unsigned int t, n;
 
   std::cin >> t;
-  for (int i = 0; i < t; ++i) {
+  for (long unsigned int i = 0; i < t; ++i) {
     std::cin >> n;
     simmetry(n);
     if (i != t - 1)
@@ -19,20 +19,20 @@ int main() {
   return 0;
 }
 
-std::vector<int> input(int size) {
-  std::vector<int> res(size);
-  for (int i = 0; i < size; ++i) {
+std::vector<long unsigned int> input(long unsigned int size) {
+  std::vector<long unsigned int> res(size);
+  for (long unsigned int i = 0; i < size; ++i) {
     std::cin >> res[i];
   }
 
   return res;
 }
 
-void simmetry(int n) {
-  std::vector<int> fence = input(n);
-  int k = 1, j = 0;
+void simmetry(long unsigned int n) {
+  std::vector<long unsigned int> fence = input(n);
+  long unsigned int k = 1, j = 0;
   if (fence[0] == n) {
-    for (int i = n - 1; i > +0; --i) {
+    for (long unsigned int i = n - 1; i > +0; --i) {
       while (fence[i] >= k) {
         if (fence[j] != i + 1) {
           std::cout << "no";
